requirements.txt
polars>=0.20.14
pandas>=2.2.2
openpyxl>=3.1.2
python-dotenv>=1.0.1
tqdm>=4.66.4
snowflake-connector-python>=3.10.0
requests>=2.31.0
pyarrow<19.0.0